requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
fdb>=2.0.2
python-dotenv>=1.0
